adjustments
when testing application in production we saw a few issues that needed adjusted. I added a line in the schedule.rb to get the mysql password it was causing import jobs to fail. Adjusted the docker-compose.yml to have the mysql password for both the mysql container and the rails container. Also needed to adjust the spec_helper.rb was having some issues with the shoulda/matchers added the require.